Refactor NgoPendingDto to NgoDto and update references
Renamed NgoPendingDto to NgoDto and moved it to the common DTOs package. Updated all service interfaces, implementations, and controller methods to use NgoDto instead of NgoPendingDto for pending NGO operations. This change improves DTO reuse and consistency across the codebase.
