+++ src/main/java/com/platform/user_service/services/Ngo/INgoPendingService.java
@@ -1,6 +1,6 @@
 package com.platform.user_service.services.Ngo;
 
-import com.platform.user_service.dtos.response.NgoPendingDto;
+import com.platform.user_service.dtos.common.NgoDto;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,5 +13,5 @@ public interface INgoPendingService {
      *
      * @return a list of NgoPendingDto representing the pending NGOs
      */
-    List<NgoPendingDto> getPending();
+    List<NgoDto> getPending();
 }
